Fix minor typo in 03_functions
Fixes inconsistent verb tense use on line 363 of 03_functions.

Before: In one case, `console.log` [had] to jump back to the `greet` function. In the other case, it jumps back to the end of the program.
After: In one case, `console.log` [has] to jump back to the `greet` function. In the other case, it jumps back to the end of the program.

Use of "has" is consistent with the rest of the paragraph, which is written in present tense.