Add 'table-layout: fixed' to book's CSS in Ch. 15
This commit adds a line -- 'table-layout:fixed' -- to the
'.background' CSS displayed in the book's Chapter 15 ('Project: A
Platform Game'). This line is required in order for the game to
function properly in Google Chrome (at least versions 38 and 39). If
the line is not included, Google Chrome's table layout algorithm can
alter the size of the table cells constituting the game's background,
resulting in buggy behavior (e.g., the player will 'hit' walls before
the player icon is actually touching a displayed wall).

The relevant line is already included in the downloadable code for the
book. This commit simply adds it, together with a brief explanation, to
the CSS displayed in the book text, so that readers who are typing
along with the book will end up with a working game.