Improve bookmarks position on mobile
Enlarge the horizontal content padding to provide more room for the
bookmarks on smaller viewports.

On smaller viewports show the bookmarks inside the boundaries of the
browser viewport (to avoid horizontal scrolling).

Provide different position to bookmarks inside blockquote elements
because the blockquote have relative position which affects the bookmark
positioning.