Clear sandbox console by replacing the whole div
Even with the original fix for #66, the sandbox didn't work for me on
macOS 10.12.4. On a hunch (the problem appeared only after clearing the
non-empty output once, so...), I've tried different methods of clearing
the output, found on StackOverflow (hey, I'm learning! Great book BTW)
instead of trying to nudge the output to reappear as in the original
fix. This seems to work for me, at least with the latest browsers on
macOS and Windows 10.