Makes comparison by identity clearer (?)
The section with an `{{index [object, identity], identity, memory, mutability}}` wasn't explicitly using the word `identity` in the explanation which made later references to "identity" confusing to me. I added in a sentence about that and also clarified that both `===` and `==` did identity comparisons in the exercise. See issue #418.

Hope that helps other readers.